requirements.txt
aiohttp==3.10.10
tiktoken==0.8.0
python-dotenv==1.0.1
pydantic==2.9.2
openai==1.51.2
tqdm==4.66.5
rank-bm25==0.2.2
jieba==0.42.1
pandas==2.2.3
faiss-cpu==1.9.0.post1
langchain==0.3.3
click==8.1.7
httpx==0.27.2
sentence-transformers==3.3.1
